Add addElement and addText command
